backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alex Simon <aleandre6579.simon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pytest = "^8.0.2"
pre-commit = "^3.6.2"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.16"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
